[Chore] 리뷰반영 및 ListLanding 컴포넌트 추가해서 context.consumer -> useContext 방식으로 변경
